backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic
langchain==0.0.350
langchain-community==0.0.10
langchain-huggingface
langchain-chroma==0.0.1
chromadb==0.4.18
openai==1.3.7
python-docx==1.1.0
PyPDF2
sentence-transformers==2.2.2
numpy==1.24.3
pandas
scikit-learn
matplotlib
seaborn
plotly
jinja2
python-pptx
openpyxl
xlrd
# Word文档解析依赖
docx2txt
unstructured==0.11.8
unstructured-inference
# PDF解析增强
PyMuPDF==1.23.8
pdfplumber
# OCR处理扫描版PDF
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
# 其他文档处理
python-magic
deepseek-ai==0.1.0
python-dotenv==1.0.0
pypdf==3.17.4
textract==1.6.5
PyMuPDF==1.23.8
opencv-python==4.8.1.78
Pillow==10.1.0
pytesseract==0.3.10
numpy==1.24.3
sqlalchemy==2.0.23
pyjwt==2.8.0
passlib==1.7.4
python-jose==3.3.0